improve buttons logic
