+++ apps/site/src/app/[locale]/simulateur/(simulator-flow)/(root)/intercalaire/_components/TransitionButtons.tsx
@@ -57,6 +57,7 @@ export default function TransitionButtons() {
           'common.previousExtended',
           'Précédent, revenir à la page précédente'
         )}
+        isClickableOnce
         disabled={isPending}
         href={SIMULATOR_PATH}
         className="h-full w-14 md:w-auto"
